[Perf][ROCm][Spec Decode] Enable 3D Split-KV Flash-Decoding for multi-query speculative verification
- Allow unified_attention to use 3D parallel Split-KV reduction kernel when q.shape[0] <= seq_threshold_3D, enabling fast verification during speculative decoding (e.g. MTP, DFlash) on ROCm.
- Fix mixed-precision dtype alignment and layer indexing for DFlash/DFlash2 drafters.
- Add VLLM_PROFILE_STEP environment variable to GPUModelRunner for synchronized step-by-step target/draft timing breakdown.

+++ vllm/v1/worker/gpu/spec_decode/dflash/speculator.py
@@ -37,7 +37,7 @@ class DFlashSpeculator(DraftModelSpeculator):
         super().__init__(vllm_config, device)
 
         self.hidden_states = torch.zeros(
-            self.max_num_tokens, self.hidden_size, dtype=self.dtype, device=device
+            self.max_num_tokens, self.hidden_size, dtype=self.draft_model_config.dtype, device=device
         )
 
         # Multimodal inputs not currently supported.
